+++ lib/components/Tabset/Tabset.d.ts
@@ -4,5 +4,6 @@ export interface ITabsetProps {
     children?: React.ReactElement<ITabProps> | Array<React.ReactElement<ITabProps>>;
     className?: string;
     onTabSelect?: (eventKey: string) => void;
+    selectedEventKey?: string;
 }
-export declare const Tabset: ({ children, className, onTabSelect }: ITabsetProps) => JSX.Element;
+export declare const Tabset: ({ children, className, onTabSelect, selectedEventKey }: ITabsetProps) => JSX.Element;